+++ gen/com/intellij/idea/plugin/hybris/impex/psi/ImpexAnyAttributeName.java
@@ -1,10 +1,6 @@
 /*
- * ----------------------------------------------------------------
- * --- WARNING: THIS FILE IS GENERATED AND WILL BE OVERWRITTEN! ---
- * ----------------------------------------------------------------
- *
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -29,5 +25,8 @@ public interface ImpexAnyAttributeName extends PsiElement {
 
   @NotNull
   List<ImpexString> getStringList();
+
+  @Nullable
+  ImpexAnyAttributeValue getAnyAttributeValue();
 
 }