+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexAnyAttributeValueImpl.java
@@ -1,10 +1,6 @@
 /*
- * ----------------------------------------------------------------
- * --- WARNING: THIS FILE IS GENERATED AND WILL BE OVERWRITTEN! ---
- * ----------------------------------------------------------------
- *
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -56,6 +52,12 @@ public class ImpexAnyAttributeValueImpl extends ImpexAttributeValueMixin impleme
   @NotNull
   public List<ImpexString> getStringList() {
     return PsiTreeUtil.getChildrenOfTypeAsList(this, ImpexString.class);
+  }
+
+  @Override
+  @Nullable
+  public ImpexAnyAttributeName getAnyAttributeName() {
+    return ImpexPsiUtil.getAnyAttributeName(this);
   }
 
 }